WIP fixing floating polygon example
