+++ examples/layer-browser/src/data-samples.js
@@ -104,19 +104,6 @@ function getRandomArbitrary(min, max) {
 // Extract simple/complex polygons arrays from geojson
 export const floatingPolygons = choropleths.features.map(choropleth =>
   choropleth.geometry.coordinates.map(coordinates =>
-    coordinates.map(coords =>
-      [...coords, getRandomArbitrary(1000, 5000), getRandomArbitrary(8000, 10000)]
-    )
-  )
-);
-
-function getRandomArbitrary(min, max) {
-  return Math.random() * (max - min) + min;
-}
-
-// Extract simple/complex polygons arrays from geojson
-export const floatingPolygons = choropleths.features.map(choropleth =>
-  choropleth.geometry.coordinates.map(coordinates =>
     coordinates.map(coords => [
       ...coords,
       getRandomArbitrary(1000, 5000),
